feat: money columns become NUMERIC (Plan 17 Slice 5, PR 2/2 — slice closed early)
The ledger and the cap are now exact at rest: `usage_event.cost`,
`pending_usage_event.cost`, `team_budget.limit_cost` and the alert outbox's
`spend`/`limit_cost` are `NUMERIC(20, 6)`, and the entities that carry them are
`Decimal`. `SUM` over the ledger is exact and order-independent, which is the
number the budget gate compares against the cap.

A `Money` column type keeps the two dialects consistent: PostgreSQL stores
NUMERIC natively, SQLite has no numeric affinity and hands back binary noise, so
the type quantizes on the way out. SQLite is a development dialect only —
production refuses it — and the round-trip is now a test on both rather than an
assumption, including the sum of 0.1 + 0.2 + 0.3 and a total that must not
depend on insertion order.

Two deliberate exclusions, both recorded in the plan:

- per-token RATE columns stay `float`. At cost scale a 0.0000005 rate would
  round to 0.000001 and corrupt the savings arithmetic, and `rate_card()`
  already converts every rate exactly via `Decimal(str(x))` when reading it, so
  the calculation is exact today. Converting them would cost another wave of
  churn for a property already held — which is why PR 3/3 is dropped and the
  slice closes here;
- the API edge stays JSON numbers, so the OpenAPI/TypeScript drift gate and the
  console are untouched.

One real defect surfaced on the way: the budget-alert webhook serializes its
payload to JSON, and `Decimal` is not JSON-serializable — the alert would have
failed to send. Caught by the existing channel tests, fixed at that boundary
with the same float-at-the-edge rule.

+++ src/litestar_gateway/domain/budget.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 
 from litestar_gateway.domain.entities import BudgetWindow
 from litestar_gateway.domain.exceptions import InvalidBudget
@@ -30,7 +31,7 @@ def validate_thresholds(thresholds: list[int]) -> list[int]:
 
 
 def crossed_thresholds(
-    *, spend: float, limit_cost: float, thresholds: list[int], fired: set[int]
+    *, spend: Decimal, limit_cost: Decimal, thresholds: list[int], fired: set[int]
 ) -> list[int]:
     """Thresholds newly crossed by `spend` against `limit_cost`, excluding any
     already in `fired` (the persisted dedup set for the current
